% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_attribute}
\alias{add_attribute}
\title{Establece los valores de un nuevo atributo específico en el conjunto de datos}
\usage{
add_attribute(.Object, attribute, values)
}
\arguments{
\item{attribute}{Nombre del atributo a establecer.}

\item{values}{Valores a asignar al atributo.}
}
\description{
Establece los valores de un nuevo atributo específico en el conjunto de datos
}
