% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_attribute_auc}
\alias{calculate_attribute_auc}
\title{Calculate the AUC of a specific attribute in the dataset relative to the class attribute.}
\usage{
calculate_attribute_auc(dataset, attribute, class_attribute)
}
\arguments{
\item{dataset}{(data.frame) A data frame containing the dataset.}

\item{attribute}{(character) The name of the attribute to calculate the AUC.}

\item{class_attribute}{(character) The name of the binary class attribute.}
}
\value{
The AUC value of the attribute.
}
\description{
Calculate the AUC of a specific attribute in the dataset relative to the class attribute.
}
\examples{
dataset <- data.frame(A = c(1, 2, 3, 4, 5), B = c(2.5, 4.5, 6.5, 8.5, 10.5), Class = c(0, 1, 1, 0, 1))
calculate_attribute_auc(dataset, "B", "Class")

}
