% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_attribute_entropy}
\alias{calculate_attribute_entropy}
\title{Calculate the entropy of a specific attribute in the dataset.}
\usage{
calculate_attribute_entropy(dataset, attribute)
}
\arguments{
\item{dataset}{(data.frame) A data frame containing the dataset.}

\item{attribute}{(character) The name of the attribute to calculate the entropy.}
}
\value{
The entropy value of the attribute.
}
\description{
Calculate the entropy of a specific attribute in the dataset.
}
\examples{
dataset <- data.frame(A = c(1, 2, 3, 4, 5), B = c("A", "B", "A", "B", "A"))
calculate_attribute_entropy(dataset, "B")

}
