% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_attribute_variance}
\alias{calculate_attribute_variance}
\title{Calculate the variance of a specific attribute in the dataset.}
\usage{
calculate_attribute_variance(dataset, attribute)
}
\arguments{
\item{dataset}{(data.frame) A data frame containing the dataset.}

\item{attribute}{(character) The name of the attribute to calculate the variance.}
}
\value{
The variance value of the attribute.
}
\description{
Calculate the variance of a specific attribute in the dataset.
}
\examples{
dataset <- data.frame(A = c(1, 2, 3, 4, 5), B = c(2.5, 4.5, 6.5, 8.5, 10.5))
calculate_attribute_variance(dataset, "B")

}
