% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_auc}
\alias{calculate_auc}
\title{Calculate the area under the ROC curve (AUC) of a continuous attribute with respect to class labels.}
\usage{
calculate_auc(attribute_values, class_labels)
}
\arguments{
\item{attribute_values}{(vector) A vector containing the values of the continuous attribute.}

\item{class_labels}{(vector) A vector containing the binary class labels (0 or 1).}
}
\value{
A list containing the true positive rate (TPR) values, false positive rate (FPR) values, and the AUC value.
}
\description{
Calculate the area under the ROC curve (AUC) of a continuous attribute with respect to class labels.
}
\examples{
attribute_values <- c(1, 2, 3, 4, 5)
class_labels <- c(0, 1, 1, 0, 1)
calculate_auc(attribute_values, class_labels)

}
