% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{calculate_categorical_mutual_information}
\alias{calculate_categorical_mutual_information}
\title{Calculate the mutual information between two categorical variables.}
\usage{
calculate_categorical_mutual_information(variable1, variable2, data)
}
\arguments{
\item{variable1}{Name of the first variable.}

\item{variable2}{Name of the second variable.}

\item{data}{DataFrame that contains the data.}
}
\value{
Mutual information between the two variables.
}
\description{
Calculate the mutual information between two categorical variables.
}
