% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{calculate_correlation}
\alias{calculate_correlation}
\title{Calculate the correlation between two variables in a dataset.}
\usage{
calculate_correlation(dataset, variable1, variable2)
}
\arguments{
\item{dataset}{An object (data frame or similar) containing the data.}

\item{variable1}{Name of the first variable.}

\item{variable2}{Name of the second variable.}
}
\value{
The correlation coefficient between the two variables.
}
\description{
Calculate the correlation between two variables in a dataset.
}
