% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{calculate_correlation_matrix}
\alias{calculate_correlation_matrix}
\title{Calculate the correlation (mutual information for categorical variables) pairwise between the variables in the dataset.}
\usage{
calculate_correlation_matrix(dataset)
}
\arguments{
\item{dataset}{An object (data frame or similar) containing the data.}
}
\value{
A data frame that contains the correlation matrix.
}
\description{
Calculate the correlation (mutual information for categorical variables) pairwise between the variables in the dataset.
}
