% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_entropy}
\alias{calculate_entropy}
\title{Calculate the entropy of a discrete attribute.}
\usage{
calculate_entropy(attribute_values)
}
\arguments{
\item{attribute_values}{(vector) A vector containing the values of the discrete attribute.}
}
\value{
The entropy value of the attribute.
}
\description{
Calculate the entropy of a discrete attribute.
}
\examples{
attribute_values <- c("A", "B", "A", "C", "B")
calculate_entropy(attribute_values)

}
