% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{calculate_numeric_correlation}
\alias{calculate_numeric_correlation}
\title{Calculate the correlation coefficient between two numeric variables.}
\usage{
calculate_numeric_correlation(values1, values2)
}
\arguments{
\item{values1}{Values of the first variable.}

\item{values2}{Values of the second variable.}
}
\value{
Correlation coefficient between the two variables.
}
\description{
Calculate the correlation coefficient between two numeric variables.
}
