% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_variance}
\alias{calculate_variance}
\title{Calculate the variance of a continuous attribute.}
\usage{
calculate_variance(attribute_values)
}
\arguments{
\item{attribute_values}{(vector) A vector containing the values of the continuous attribute.}
}
\value{
The variance value of the attribute.
}
\description{
Calculate the variance of a continuous attribute.
}
\examples{
attribute_values <- c(1, 2, 3, 4, 5)
calculate_variance(attribute_values)

}
