% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{class}
\name{Dataset-class}
\alias{Dataset-class}
\alias{Dataset}
\title{Clase Dataset para trabajar con conjuntos de datos}
\description{
Clase Dataset para trabajar con conjuntos de datos
}
\section{Slots}{

\describe{
\item{\code{data}}{DataFrame de pandas que contiene los datos del conjunto de datos}
}}

