% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{discretize}
\alias{discretize}
\title{Discretize input values \code{x} into intervals defined by cut points.}
\usage{
discretize(x, cut_points)
}
\arguments{
\item{x}{(vector) A vector of numeric values to discretize.}

\item{cut_points}{(vector) A vector of cut points that define the intervals.}
}
\value{
A list with two elements:
\itemize{
\item \code{x_discretized}: A list of discretized values represented as strings in the form "I\if{html}{\out{<n>}}",
where \if{html}{\out{<n>}} is the interval index.
\item \code{cut_points}: The original list of cut points, unchanged.
}
}
\description{
Discretize input values \code{x} into intervals defined by cut points.
}
\examples{
x <- c(1, 2, 3, 4, 5)
cut_points <- c(2, 4)
discretize(x, cut_points)

}
