% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{discretize_attribute}
\alias{discretize_attribute}
\title{Discretize an attribute using the specified method.}
\usage{
discretize_attribute(attribute, num_bins, method)
}
\arguments{
\item{attribute}{(vector) An attribute represented as a vector.}

\item{num_bins}{(numeric or vector) The number of bins to discretize each attribute of the dataset.}

\item{method}{(character) The discretization method to use. Can be 'equal_width' or 'equal_frequency'.}
}
\value{
A list with two elements:
\itemize{
\item \code{discretized_attribute}: A new attribute with discretized values. A list of discretized values
represented as strings in the form "I\if{html}{\out{<n>}}", where \if{html}{\out{<n>}} is the interval index.
\item \code{cut_points}: The list of cut points used for discretization.
}
}
\description{
Discretize an attribute using the specified method.
}
\examples{
attribute <- c(1, 2, 3, 4, 5)
num_bins <- 3
method <- "equal_width"
discretize_attribute(attribute, num_bins, method)

}
