% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{discretize_dataset}
\alias{discretize_dataset}
\title{Discretize a dataset using the specified method.}
\usage{
discretize_dataset(dataset, num_bins, method)
}
\arguments{
\item{dataset}{(list) A dataset represented as a list of vectors.}

\item{num_bins}{(numeric) The number of bins to discretize each attribute of the dataset.}

\item{method}{(character) The discretization method to use. Can be 'equal_width' or 'equal_frequency'.}
}
\value{
A list representing the discretized dataset.
}
\description{
Discretize a dataset using the specified method.
}
\examples{
dataset <- list(
  attribute1 = c(1, 2, 3, 4, 5),
  attribute2 = c(1.5, 2.5, 3.5, 4.5, 5.5),
  attribute3 = c("A", "B", "C", "D", "E")
)
num_bins <- 3
method <- "equal_width"
discretize_dataset(dataset, num_bins, method)

}
