% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{discretizeEF}
\alias{discretizeEF}
\title{Discretize an input attribute using Equal Frequency discretization.}
\usage{
discretizeEF(x, num_bins)
}
\arguments{
\item{x}{(vector) A vector of numeric values.}

\item{num_bins}{(integer) The number of intervals to discretize \code{x} into.}
}
\value{
A list with two elements:
\itemize{
\item \code{x_discretized}: A list of discretized values represented as strings in the form "I\if{html}{\out{<n>}}",
where \if{html}{\out{<n>}} is the interval index.
\item \code{cut_points}: A list of cut points between intervals.
}
}
\description{
Discretize an input attribute \code{x} into \code{num_bins} intervals using Equal Frequency discretization.
}
\examples{
x <- c(1, 2, 3, 4, 5)
num_bins <- 3
discretizeEF(x, num_bins)

}
