% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{empty}
\alias{empty}
\title{Comprueba si el objeto Dataset está vacío}
\usage{
empty(.Object)
}
\value{
TRUE si el objeto Dataset está vacío, FALSE de lo contrario.
}
\description{
Comprueba si el objeto Dataset está vacío
}
