% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_variables}
\alias{filter_variables}
\title{Filter the variables of the dataset based on the specified metric and threshold criteria.}
\usage{
filter_variables(dataset, metric, threshold, operator, class_attribute = NULL)
}
\arguments{
\item{dataset}{(data.frame) A data frame containing the dataset.}

\item{metric}{(character) The metric to use for filtering the variables.
Can be 'entropy', 'variance', or 'auc'.}

\item{threshold}{(numeric) The threshold value to compare with the metric.}

\item{operator}{(character) The operator to use for comparing the metric with the threshold.
Can be '>', '<', or '=='.}

\item{class_attribute}{(character) Optional name of the class attribute used for calculating AUC.}
}
\value{
A new data frame containing the filtered variables.
}
\description{
Filter the variables of the dataset based on the specified metric and threshold criteria.
}
\examples{
dataset <- data.frame(A = c(1, 2, 3, 4, 5), B = c(2.5, 4.5, 6.5, 8.5, 10.5), Class = c(0, 1, 1, 0, 1))
filter_variables(dataset, "variance", 5, ">")

}
