% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_attribute}
\alias{get_attribute}
\title{Obtiene los valores de un atributo específico del conjunto de datos}
\usage{
get_attribute(.Object, attribute)
}
\arguments{
\item{attribute}{Nombre del atributo deseado.}
}
\value{
Serie de pandas que contiene los valores del atributo especificado.
}
\description{
Obtiene los valores de un atributo específico del conjunto de datos
}
