% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_attributes}
\alias{get_attributes}
\title{Obtiene una lista de los nombres de los atributos en el conjunto de datos}
\usage{
get_attributes(.Object)
}
\value{
Lista de strings que representa los nombres de los atributos.
}
\description{
Obtiene una lista de los nombres de los atributos en el conjunto de datos
}
