% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_column}
\alias{get_column}
\title{Obtiene una columna específica del conjunto de datos}
\usage{
get_column(.Object, i)
}
\arguments{
\item{i}{Índice de la columna deseada.}
}
\value{
Serie de pandas que contiene los valores de la columna especificada.
}
\description{
Obtiene una columna específica del conjunto de datos
}
