% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_data}
\alias{get_data}
\title{Obtiene el DataFrame de pandas que representa los datos del conjunto de datos}
\usage{
get_data(.Object)
}
\value{
DataFrame de pandas que contiene los datos del conjunto de datos.
}
\description{
Obtiene el DataFrame de pandas que representa los datos del conjunto de datos
}
