% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{initialize}
\alias{initialize}
\title{Inicializa un objeto Dataset}
\usage{
initialize(.Object, data = NULL)
}
\arguments{
\item{data}{DataFrame de pandas opcional que contiene los datos del conjunto de datos.}
}
\description{
Inicializa un objeto Dataset
}
