% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{normalize_dataset}
\alias{normalize_dataset}
\title{Normalize all numeric variables in a dataset.}
\usage{
normalize_dataset(dataset)
}
\arguments{
\item{dataset}{A dataset (data frame or similar).}
}
\value{
A new dataset with normalized numeric variables.
}
\description{
Normalize all numeric variables in a dataset.
}
