% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{normalize_variable}
\alias{normalize_variable}
\title{Normalize a numeric variable using min-max normalization.}
\usage{
normalize_variable(x)
}
\arguments{
\item{x}{A vector or array of numeric values.}
}
\value{
A vector with the normalized values of 'x'.
}
\description{
Normalize a numeric variable using min-max normalization.
}
\examples{
x <- c(1, 2, 3, 4, 5)
normalize_variable(x)

}
