% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{num_attributes}
\alias{num_attributes}
\title{Obtiene el número de atributos en el conjunto de datos}
\usage{
num_attributes(.Object)
}
\value{
Entero que representa el número de atributos.
}
\description{
Obtiene el número de atributos en el conjunto de datos
}
