% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{read_data}
\alias{read_data}
\title{Lee los datos desde un archivo y los carga en el objeto Dataset}
\usage{
read_data(.Object, file_path)
}
\arguments{
\item{file_path}{Ruta del archivo que contiene los datos.}
}
\value{
El objeto Dataset con los datos cargados.
}
\description{
Lee los datos desde un archivo y los carga en el objeto Dataset
}
