% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{standardize_dataset}
\alias{standardize_dataset}
\title{Standardize all numeric variables in a dataset.}
\usage{
standardize_dataset(dataset)
}
\arguments{
\item{dataset}{A dataset (data frame or similar).}
}
\value{
A new dataset with standardized numeric variables.
}
\description{
Standardize all numeric variables in a dataset.
}
