% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{standardize_variable}
\alias{standardize_variable}
\title{Standardize a numeric variable using z-score standardization.}
\usage{
standardize_variable(x)
}
\arguments{
\item{x}{A vector or array of numeric values.}
}
\value{
A vector with the standardized values of 'x'.
}
\description{
Standardize a numeric variable using z-score standardization.
}
\examples{
x <- c(1, 2, 3, 4, 5)
standardize_variable(x)

}
