% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{write_data}
\alias{write_data}
\title{Escribe los datos del objeto Dataset en un archivo}
\usage{
write_data(.Object, file_path)
}
\arguments{
\item{file_path}{Ruta del archivo donde se guardarán los datos.}
}
\description{
Escribe los datos del objeto Dataset en un archivo
}
